feat(collections): add support of iterable in intersect
